requirements.txt
yt_dlp
openai-whisper
torch 
transformers
nltk
flask
scikit-learn
flask_mysqldb